reorganize rules tests into type: find create etc
