+++ packages/usage/__tests__/integration/rules/rules.allowed-blocked.test.ts
@@ -1,10 +1,10 @@
 import { beforeEach, expect, it } from '@jest/globals';
-import { mockFetch } from '../__mocks__/fetch.mock';
-import bridg from '../generated/bridg';
-import { Blog, User } from '../generated/prisma';
-import { deleteDbData, seedDbData } from '../utils/prisma.test-util';
-import { queryFails, querySucceeds } from '../utils/query.test-util';
-import { setRules } from '../utils/rules.test-util';
+import { mockFetch } from '../../__mocks__/fetch.mock';
+import bridg from '../../generated/bridg';
+import { Blog, User } from '../../generated/prisma';
+import { deleteDbData, seedDbData } from '../../utils/prisma.test-util';
+import { queryFails, querySucceeds } from '../../utils/query.test-util';
+import { setRules } from '../../utils/rules.test-util';
 
 global.fetch = mockFetch;
